SIGBUS and SIGSEGV workaround
We noticed in our game that the Texture2D.Destroy method was causing memory exception crashes. Something seems to be trying to access the memory address after unity destroyed the dynamic texture. We moved the unload method to the finalizer, and the problem seems to be resolved. What could be trying to access the memory after you asked to be destroyed?